Backend/Auth/files/requirements.txt
fastapi
uvicorn
python-dotenv
pymongo
passlib
python-jose
longtrainer
pydantic[email]
